checkmarx refactoring
